Move the person interface tests to the shared examples folder
