+++ tests/shared_examples/person_interface_test.rb
